MMFNEXUS

Documentation for MMFNEXUS.

MMFNEXUS.FourierFilter

MMFNEXUS.FourierFilter

MMFNEXUS.SimBox

MMFNEXUS.compute_signatures!

MMFNEXUS.fft_field!

MMFNEXUS.fila_wall_field

MMFNEXUS.insignificant_fila_wall_remover

MMFNEXUS.insignificant_node_remover

MMFNEXUS.multiscale_signature_max!

MMFNEXUS.node_field

MMFNEXUS.node_threshold_fraction

MMFNEXUS.signatures_hessian!

MMFNEXUS.FourierFilter — Type

(filt::FourierFilter)(ρ::Array{<:Real,3}, sim_box::SimBox, R::Real; log::Bool=false)

Apply the Gaussian Fourier filter to a real 3D density field ρ with smoothing scale R. Reuses internal memory and performs FFT in-place.

Returns the filtered density field as a new array.

MMFNEXUS.FourierFilter — Type

FourierFilter(sim_box::SimBox)

Precomputes the k-grid and allocates reusable work arrays. Call the object like a function to filter a density field in-place.

MMFNEXUS.SimBox — Type

SimBox(N,L,M)

Holds the simulation box parameters:
N : Number of grid points per dimension
L : Box size in cMpc/h
M : Mass in Msun/h contained in the box
Generates also additional useful parameters:
ρ_mean : Mean density in Msun/h/(cMpc/h)^3
VoxelVolume : Volume of a voxel in (cMpc/h)^3

MMFNEXUS.compute_signatures! — Method

Compute signatures from Hessian eigenvalues.

MMFNEXUS.fft_field! — Method

fft_field!(filt::FourierFilter, ρ::Array{<:Real,3})

Fourier transform a real 3D field ρ, returning the Fourier-space field.

MMFNEXUS.fila_wall_field — Method

fila_wall_field(sim_box::SimBox, ρ::Array{<:Real,3}, signatures::Array{<:Real,3}, clean_field, min_vol::Real; s_low=1e-5, s_high=1e-2, stepsize=0.1)

Generate clean filament/wall field based on significant filaments/walls above optimal signature threshold.

MMFNEXUS.insignificant_fila_wall_remover — Method

insignificant_fila_wall_remover(sim_box::SimBox, signatures::Array{<:Real,3}, S_th::Real, min_vol::Real)

Remove filament/wall objects below minimum volume from the signature field at given signature threshold S_th.
Returns a boolean array indicating significant filaments/walls.

MMFNEXUS.insignificant_node_remover — Method

insignificant_node_remover(sim_box::SimBox, ρ::Array{<:Real,3}, signature::Array{<:Real,3}, S_th::Real, min_node_mass::Real)

Remove node objects below minimum mass from the signature field at given signature threshold S_th.
Returns a boolean array indicating significant nodes.

MMFNEXUS.multiscale_signature_max! — Method

multiscale_signature_max!(filt::FourierFilter, sim_box::SimBox, ρ::Array{<:Real,3}, R0::Real; mode::Symbol, scalespec)
Compute maximum signatures across multiple scales specified by the user.

MMFNEXUS.node_field — Method

node_field(sim_box::SimBox, ρ::Array{<:Real,3}, signatures::Array{<:Real,3}, min_node_mass::Real, Δ::Real; s_low=1e-5, s_high=9e-1, fraction_target=0.5)

Generate clean node field based on significant nodes above optimal signature threshold.

MMFNEXUS.node_threshold_fraction — Method

node_threshold_fraction(sim_box::SimBox, ρ::Array{<:Real,3}, signature::Array{<:Real,3}, S_th::Real, min_node_mass::Real, Δ::Real)

Compute the fraction of identified nodes above overdensity Δ at signature threshold S_th.
This computation is used to find the optimal signature threshold for node identification.

MMFNEXUS.signatures_hessian! — Method

signatures_hessian!(filt::FourierFilter, sim_box::SimBox, ρ::Array{<:Real,3}, R::Real; mode::Symbol)

Compute Hessian eigenvalue signatures (node, filament, wall) at smoothing scale R.